fix: delegate some of span model parts into aggregator
