+++ internal/sampling/samplingstrategy/adaptive/aggregator_test.go
@@ -5,6 +5,7 @@ package adaptive
 
 import (
 	"net/http"
+	"strconv"
 	"testing"
 	"time"
 
@@ -185,4 +186,124 @@ func TestRecordThroughputFunc(t *testing.T) {
 	}
 	a.HandleRootSpan(span)
 	assert.EqualValues(t, 1, a.(*aggregator).currentThroughput["A"][http.MethodGet].Count)
+}
+
+func TestGetSamplerParams(t *testing.T) {
+	logger := zap.NewNop()
+	tests := []struct {
+		tags          model.KeyValues
+		expectedType  model.SamplerType
+		expectedParam float64
+	}{
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "probabilistic"),
+				model.String("sampler.param", "1e-05"),
+			},
+			expectedType:  model.SamplerTypeProbabilistic,
+			expectedParam: 0.00001,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "probabilistic"),
+				model.Float64("sampler.param", 0.10404450002098709),
+			},
+			expectedType:  model.SamplerTypeProbabilistic,
+			expectedParam: 0.10404450002098709,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "probabilistic"),
+				model.String("sampler.param", "0.10404450002098709"),
+			},
+			expectedType:  model.SamplerTypeProbabilistic,
+			expectedParam: 0.10404450002098709,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "probabilistic"),
+				model.Int64("sampler.param", 1),
+			},
+			expectedType:  model.SamplerTypeProbabilistic,
+			expectedParam: 1.0,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "ratelimiting"),
+				model.String("sampler.param", "1"),
+			},
+			expectedType:  model.SamplerTypeRateLimiting,
+			expectedParam: 1,
+		},
+		{
+			tags: model.KeyValues{
+				model.Float64("sampler.type", 1.5),
+			},
+			expectedType:  model.SamplerTypeUnrecognized,
+			expectedParam: 0,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "probabilistic"),
+			},
+			expectedType:  model.SamplerTypeUnrecognized,
+			expectedParam: 0,
+		},
+		{
+			tags:          model.KeyValues{},
+			expectedType:  model.SamplerTypeUnrecognized,
+			expectedParam: 0,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "lowerbound"),
+				model.String("sampler.param", "1"),
+			},
+			expectedType:  model.SamplerTypeLowerBound,
+			expectedParam: 1,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "lowerbound"),
+				model.Int64("sampler.param", 1),
+			},
+			expectedType:  model.SamplerTypeLowerBound,
+			expectedParam: 1,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "lowerbound"),
+				model.Float64("sampler.param", 0.5),
+			},
+			expectedType:  model.SamplerTypeLowerBound,
+			expectedParam: 0.5,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "lowerbound"),
+				model.String("sampler.param", "not_a_number"),
+			},
+			expectedType:  model.SamplerTypeUnrecognized,
+			expectedParam: 0,
+		},
+		{
+			tags: model.KeyValues{
+				model.String("sampler.type", "not_a_type"),
+				model.String("sampler.param", "not_a_number"),
+			},
+			expectedType:  model.SamplerTypeUnrecognized,
+			expectedParam: 0,
+		},
+	}
+
+	for i, test := range tests {
+		tt := test
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			span := &model.Span{}
+			span.Tags = tt.tags
+			actualType, actualParam := GetSamplerParams(span, logger)
+			assert.Equal(t, tt.expectedType, actualType)
+			assert.InDelta(t, tt.expectedParam, actualParam, 0.01)
+		})
+	}
 }
