+++ model/span.go
@@ -7,9 +7,6 @@ package model
 import (
 	"encoding/gob"
 	"io"
-	"strconv"
-
-	"go.uber.org/zap"
 )
 
 type SamplerType int
@@ -141,37 +138,4 @@ func (s *Span) ReplaceParentID(newParentID SpanID) {
 		}
 	}
 	s.References = MaybeAddParentSpanID(s.TraceID, newParentID, s.References)
-}
-
-// GetSamplerParams returns the sampler.type and sampler.param value if they are valid.
-func (s *Span) GetSamplerParams(logger *zap.Logger) (SamplerType, float64) {
-	samplerType := s.GetSamplerType()
-	if samplerType == SamplerTypeUnrecognized {
-		return SamplerTypeUnrecognized, 0
-	}
-	tag, ok := KeyValues(s.Tags).FindByKey(SamplerParamKey)
-	if !ok {
-		return SamplerTypeUnrecognized, 0
-	}
-	samplerParam, err := samplerParamToFloat(tag)
-	if err != nil {
-		logger.
-			With(zap.String("traceID", s.TraceID.String())).
-			With(zap.String("spanID", s.SpanID.String())).
-			Warn("sampler.param tag is not a number", zap.Any("tag", tag))
-		return SamplerTypeUnrecognized, 0
-	}
-	return samplerType, samplerParam
-}
-
-func samplerParamToFloat(samplerParamTag KeyValue) (float64, error) {
-	// The param could be represented as a string, an int, or a float
-	switch samplerParamTag.VType {
-	case Float64Type:
-		return samplerParamTag.Float64(), nil
-	case Int64Type:
-		return float64(samplerParamTag.Int64()), nil
-	default:
-		return strconv.ParseFloat(samplerParamTag.AsString(), 64)
-	}
 }
